Move SPIRV passes from `Passes.td` -> `SPIRV/Passes.td`.
